#152 К отчётам добавлен номер билда
